+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/ComboBoxSelectionDialogBuilder.java
@@ -27,6 +27,13 @@ import java.util.Optional;
 public interface ComboBoxSelectionDialogBuilder<T> extends SelectionDialogBuilder<T, ComboBoxSelectionDialogBuilder<T>> {
 
 	/**
+	 * @param defaultSelection the item selected by default
+	 * @return this builder instance
+	 * @throws IllegalArgumentException in case the selection values do not contain the default selection item
+	 */
+	ComboBoxSelectionDialogBuilder<T> defaultSelection(T defaultSelection);
+
+	/**
 	 * @return the selected value, {@link Optional#empty()} if none was selected
 	 * @throws is.codion.common.model.CancelException in case the user cancelled
 	 */