+++ changelog.md
@@ -37,6 +37,7 @@ Codion Change Log
 - ControlPanelBuilder, Builder suffix removed from a few methods.
 - DefaultButtonPanelBuilder.enableTransferFocusOnEnter() bug fixed, was restricted to JButton.
 - ButtonPanelBuilder.fixedButtonSize() and buttonGroup() added.
+- SelectionDialogBuilder.defaultSelection() moved to sub-interfaces.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.